+++ src/main/java/com/ezequias/apiestudo/dto/RepresentanteDTO.java
@@ -11,14 +11,14 @@ public class RepresentanteDTO implements Serializable {
 	
 	private Integer id ;
 	private String nome;
-	private Integer metaId;
+	private Long metaId;
 	@JsonFormat(pattern="dd/MM/yyyy HH:mm")
 	private Date nascimento;
 	
 	public RepresentanteDTO() {	
 	}
 
-	public RepresentanteDTO(Integer id, String nome, Date nascimento, Integer metaId) {
+	public RepresentanteDTO(Integer id, String nome, Date nascimento, Long metaId) {
 		super();
 		this.id = id;
 		this.nome = nome;
@@ -50,11 +50,11 @@ public class RepresentanteDTO implements Serializable {
 		this.nascimento = nascimento;
 	}
 
-	public Integer getMetaId() {
+	public Long getMetaId() {
 		return metaId;
 	}
 
-	public void setMetaId(Integer metaId) {
+	public void setMetaId(Long metaId) {
 		this.metaId = metaId;
 	}
